Fix release Gradle task not compatible with config cache (#12728)
